# apps/runner/pkg/api/docs/swagger.yaml
definitions:
...
  ErrorResponse:
    description: Error response
    properties:
      code:
        example: BAD_REQUEST
        type: string
      message:
        example: Bad request
        type: string
      method:
        example: GET
        type: string
      path:
        example: /api/resource
        type: string
      statusCode:
        example: 400
        type: integer
      timestamp:
        example: '2023-01-01T12:00:00Z'
        type: string
    required:
      - message
      - path
      - statusCode
      - timestamp
    type: object
...
  generationstop.ExpectedGeneration:
    properties:
      containerCreatedAt:
        type: string
      containerId:
        type: string
      executionStartedAt:
        type: string
      restartCount:
        type: integer
    required:
      - containerCreatedAt
      - containerId
      - executionStartedAt
      - restartCount
    type: object
...
  workingcopy.CaptureAuthorityArtifact:
    properties:
      digest:
        type: string
      ref:
        type: string
    required:
      - digest
      - ref
    type: object
...
  workingcopy.CaptureComponent:
    properties:
      helper:
        $ref: '#/definitions/workingcopy.CaptureAuthorityArtifact'
      protocol:
        $ref: '#/definitions/workingcopy.CaptureAuthorityArtifact'
      roleRef:
        type: string
    required:
      - helper
      - protocol
      - roleRef
    type: object
...
  workingcopy.SandboxFileDeleteReceipt:
    properties:
      captureId:
        type: string
        x-nullable: true
      outcome:
        enum:
          - deleted
          - already_absent
        type: string
    required:
      - captureId
      - outcome
    type: object
  workingcopy.SandboxFileDeleteRequest:
    properties:
      operationId:
        type: string
      organizationId:
        type: string
      path:
        type: string
      receipt:
        $ref: '#/definitions/workingcopy.SandboxFileReceipt'
    type: object
  workingcopy.SandboxFileObservation:
    properties:
      receipt:
        $ref: '#/definitions/workingcopy.SandboxFileReceipt'
      status:
        enum:
          - absent
          - pending
          - complete
          - retired
        type: string
    required:
      - status
    type: object
  workingcopy.SandboxFileObserveRequest:
    properties:
      operationId:
        type: string
      organizationId:
        type: string
      path:
        type: string
    required:
      - operationId
      - organizationId
    type: object
  workingcopy.SandboxFileReadRequest:
    properties:
      maximumBytes:
        type: integer
      offset:
        type: integer
      receipt:
        $ref: '#/definitions/workingcopy.SandboxFileReceipt'
    required:
      - maximumBytes
      - offset
      - receipt
    type: object
  workingcopy.SandboxFileReadResponse:
    properties:
      byteLength:
        type: integer
      bytesBase64:
        type: string
      captureId:
        type: string
      eof:
        type: boolean
      offset:
        type: integer
      sha256:
        type: string
      totalByteLength:
        type: integer
    required:
      - byteLength
      - bytesBase64
      - captureId
      - eof
      - offset
      - sha256
      - totalByteLength
    type: object
  workingcopy.SandboxFileReceipt:
    properties:
      captureId:
        type: string
      capturedAt:
        type: string
      component:
        $ref: '#/definitions/workingcopy.CaptureComponent'
      contract:
        type: string
      generation:
        $ref: '#/definitions/generationstop.ExpectedGeneration'
      operationId:
        type: string
      organizationId:
        type: string
      path:
        type: string
      sandboxId:
        type: string
      sha256:
        type: string
      totalByteLength:
        type: integer
    required:
      - captureId
      - capturedAt
      - component
      - contract
      - generation
      - operationId
      - organizationId
      - path
      - sandboxId
      - sha256
      - totalByteLength
    type: object
  workingcopy.SandboxFileRequest:
    properties:
      operationId:
        type: string
      organizationId:
        type: string
      path:
        type: string
    required:
      - operationId
      - organizationId
      - path
    type: object
...
info:
  contact: {}
  description: Daytona Runner API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Daytona Runner API
  version: v0.0.0-dev
paths:
...
  /sandboxes/{sandboxId}/working-copy-captures/sandbox-files:
    post:
      consumes:
        - application/json
      description: Provider organization authority is supplied by the authenticated
        API, not a Product Run. File capture admits only work/outputs regular files
        and requires native descriptor cloning. Component measurement is not runtime
        qualification.
      operationId: CaptureSandboxFile
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Native file capture operation
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileReceipt'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Capture an immutable native sandbox file once
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/sandbox-files/delete:
    post:
      consumes:
        - application/json
      description: Provider organization authority is supplied by the authenticated
        API, not a Product Run. File capture admits only work/outputs regular files
        and requires native descriptor cloning. Component measurement is not runtime
        qualification.
      operationId: DeleteSandboxFile
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Native file capture operation
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileDeleteRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileDeleteReceipt'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Retire a native file operation and release its private bytes
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/sandbox-files/observe:
    post:
      consumes:
        - application/json
      description: Provider organization authority is supplied by the authenticated
        API, not a Product Run. File capture admits only work/outputs regular files
        and requires native descriptor cloning. Component measurement is not runtime
        qualification.
      operationId: ObserveSandboxFile
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Native file capture operation
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileObserveRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileObservation'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Observe a native sandbox file operation without source effects
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/sandbox-files/read:
    post:
      consumes:
        - application/json
      description: Provider organization authority is supplied by the authenticated
        API, not a Product Run. File capture admits only work/outputs regular files
        and requires native descriptor cloning. Component measurement is not runtime
        qualification.
      operationId: ReadSandboxFile
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Native file capture operation
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileReadRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileReadResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Read a bounded immutable native file capture
      tags:
        - sandbox
...
security:
  - Bearer: []
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and an API token.
    in: header
    name: Authorization
    type: apiKey
swagger: '2.0'